+++ site/app/templates/application.hbs
@@ -30,21 +30,21 @@
               @iconType="arrowDown"
               {{on "click" (set this "themePopover" true)}}
             >
-              {{this.currentTheme.name}}
+              {{this.themeManager.currentTheme.name}}
             </EuiButton>
           </:button>
           <:content>
             <div class="euiContextMenuPanel" tabindex="-1">
-              {{#each this.themes as |theme|}}
+              {{#each this.themeManager.themes as |theme|}}
                 <button
                   type="button"
                   class="euiContextMenuItem euiContextMenuItem--small"
-                  {{on "click" (fn this.setTheme theme)}}
+                  {{on "click" (fn this.themeManager.changeTheme theme)}}
                 >
                   <span class="euiContextMenu__itemLayout">
                     <EuiIcon
                       @type={{if
-                        (eq this.currentTheme.key theme.key)
+                        (eq this.themeManager.currentTheme.key theme.key)
                         "check"
                         "empty"
                       }}
